refactor(clock): extract waybar_time to util/waybar_time.hpp
